requirements.txt
# Core Agent
langgraph==0.2.34
langchain==0.3.7
langchain-ollama==0.2.0
langchain-core==0.3.15
langchain-community==0.3.5

# Ollama
ollama==0.4.1

# Google APIs (for Calendar and Tasks)
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

google-api-python-client==2.149.0

# MCP
mcp==1.1.2

# Observability
langfuse==2.51.3

# UI
streamlit==1.39.0
streamlit-option-menu==0.4.0
plotly==5.24.1

# Utilities
python-dotenv==1.0.0
pydantic==2.9.2
httpx==0.27.2
rich==13.9.4
python-dateutil==2.9.0
pytz==2024.2
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12